requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
lunar_python
